moved RPI updates into parallel threads
moved RPI updates into parallel threads. This accellerates the updates from ~ 40 secs to ~ 3 secs for 20 RPI. and if there is an RPI that is offline it will not stop updating the other RPI. For the user it has the advantage that the regular main plugin thread is not blocked by waiting for RPI to answer and opening menus etc will not time-out in these situations.

Also fixed some smaller bugs